Vantiv: Add authorize (credit card) request tests
Add tests to exercise the request data creation for the `authorize`
action with the `credit_card` payment method. The basic request is
tested for nodes that are created as well as nodes that are expected to
be left out of the data. Specific options that result in new/optional
nodes are tested explicity.

The `<pos>` node can be created in a variety of situations but the
internals are not changed - the values actually come from constants - so
there is a test added in the `test_xml__*` section to make sure the node
is created as expected.