Vantiv: Add store (credit card) request tests
Add tests to exercise the request data creation for the `store` action
with the `credit_card` payment method. The basic request is tested for
nodes that are created as well as nodes that are expected to be left
out. Specific options that result in new/optional nodes are explicitly
tested.