Vantiv: Add void request tests
Add tests to exercise the request data creation for the `void` action
with various authorization strings. The root node for this action
changes if the void is for an authorization (`authReversal`) versus
other actions (`capture`, `purchase`, `refund`). The request is tested
under each of these contexts for required nodes as well as those that
should not be included.