Vantiv: Add initial eProtect Registration support
Adds the `Registration` object/class to represent an eProtect
Registration Id. Since the registration needs to have an expiration date
provided in most of the transactions, the object provides a way to test
what type we're given, provide extra information, and clarify existing
usage. The existing support for `paypage` (now Vantiv eProtect) was
using a String passed in to `store` - this has been updated to use the
new class.

Tests are added for the following supported actions:
  * `authorize`
  * `purchase`
  * `refund`