Vantiv: Add missing assertions
The `auth_assertions` method which is shared across multiple tests
was lacking assertions for the response code and the auth code.

The auth code assertion was added for test 5, but it needs to be added
for some of the other tests using the `auth_assertions` method.
This will be done in a future cleanup.