Vantiv: Refactor request building
Refactors the request building to use a series of builder classes - one
for each payment method type. Each builder inherits from
`RequestBuilder` and implements only those methods that it supports. For
instance, the `AuthorizationRequestBuilder` (used for any actions called
on an `Authorization` object) only supports `capture`, `refund` , and
`void`, while `CheckRequestBuilder` supports `purchase`, `refund`, and
`store`.

The gateway sets up a small registry of builders during initialization
and those builders are delegated to inside each gateway action. The
requests are built and sent to Vantiv in the normal way.

This elaborate construction is a by product of the way in which Vantiv
expects the XML to be formed for the different kinds of requests. It
makes it explicit and easy to find where a chose action/method is
implemented.

Common xml helpers are in the `RequestBuilderHelpers` module for
organization. Common functionality used by the gateway is implemented in
private methods on the gateway itself.