Vantiv: Reorganize unit tests to use sub tests
Uses TestUnit sub tests to organize the tests. The sub test classes
primarily match the internal classes in the gateway - so each internal
object has a sub test class with tests. There are some response parsing
tests that are organized by action type (purchase, refund, void, etc)
which use sub tests to group them as well.

Helps when working on tests for specific areas of the gateway since the
internals are now organized by payment type and then the actions that
are supported for each type.